move contract logic into seprate file
